+++ modules/client/examples/01-client-erc20/03-create-proposal.ts
@@ -7,6 +7,7 @@ import {
   ContextPlugin,
   ICreateProposalParams,
   ProposalCreationSteps,
+  ProposalMetadata,
   VoteValues,
 } from "@aragon/sdk-client";
 import { contextParams } from "../00-client/00-context";
@@ -18,27 +19,31 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 // Create an ERC20 client
 const client = new ClientErc20(contextPlugin);
 
+const metadata:ProposalMetadata= {
+  title: "Test Proposal",
+  summary: "This is a short description",
+  description: "This is a long description",
+  resources: [
+    {
+      name: "Discord",
+      url: "https://discord.com/...",
+    },
+    {
+      name: "Website",
+      url: "https://website...",
+    },
+  ],
+  media: {
+    logo: "https://...",
+    header: "https://...",
+  },
+}
+
+const cid = await client.methods.pinMetadata(metadata)
+
 const proposalParams: ICreateProposalParams = {
   pluginAddress: "0x1234567890123456789012345678901234567890",
-  metadata: {
-    title: "Test Proposal",
-    summary: "This is a short description",
-    description: "This is a long description",
-    resources: [
-      {
-        name: "Discord",
-        url: "https://discord.com/...",
-      },
-      {
-        name: "Website",
-        url: "https://website...",
-      },
-    ],
-    media: {
-      logo: "https://...",
-      header: "https://...",
-    },
-  },
+  metadata:cid,
   actions: [],
   startDate: new Date(),
   endDate: new Date(),