+++ modules/client/examples/04-decoders/05-update-metadata.ts
@@ -9,7 +9,9 @@ const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
 const data: Uint8Array = new Uint8Array([12, 56]);
 
-const params: IMetadata = await client.decoding.updateMetadataAction(data);
+const cid: string = await client.decoding.updateMetadataAction(data);
+
+const params = await client.methods.fetchMetadata(cid);
 
 console.log(params);
 /*