+++ modules/client/src/erc20/interfaces.ts
@@ -16,6 +16,7 @@ import {
   ProposalBase,
   ProposalCreationStepValue,
   ProposalListItemBase,
+  ProposalMetadata,
   SubgraphAction,
   SubgraphProposalBase,
   SubgraphVoterListItemBase,
@@ -29,6 +30,7 @@ export interface IClientErc20Methods extends IClientCore {
   createProposal: (
     params: ICreateProposalParams,
   ) => AsyncGenerator<ProposalCreationStepValue>;
+  pinMetadata: (params: ProposalMetadata) => Promise<string>;
   voteProposal: (
     params: IVoteProposalParams,
   ) => AsyncGenerator<VoteProposalStepValue>;